apps/requirements.txt
streamlit>=1.32
requests>=2.31
pynacl>=1.5.0
